moved chip folder to common folder
