+++ src/common/ChipCell/chipCell.scss
@@ -27,9 +27,5 @@
         max-width: 195px;
       }
     }
-
-    .showChips .chip-block-hidden {
-      display: flex !important;
-    }
   }
 }
